move server to a discrete file
